+++ release/components/slider/slider.component.js
@@ -15,6 +15,7 @@ var SLIDER_VALUE_ACCESSOR = {
     useExisting: forwardRef(function () { return SliderComponent; }),
     multi: true
 };
+var edge = window.navigator.userAgent.indexOf("Edge") > -1;
 var SliderComponent = /** @class */ (function () {
     function SliderComponent() {
         this.id = "range-" + ++nextId;
@@ -23,12 +24,14 @@ var SliderComponent = /** @class */ (function () {
         this.step = 1;
         this.orientation = 'horizontal';
         this.filled = false;
-        // Not supported in all
-        // browers see polyfill
-        // http://leaverou.github.io/multirange/
         this.multiple = false;
+        this.disabled = false;
         this.showTicks = false;
-        this.count = [];
+        this._values = [0];
+        this._percents = [0];
+        this._thumbs = [];
+        this._ticks = [];
+        this._active = [];
         this.change = new EventEmitter();
         this.onTouchedCallback = function () {
             // placeholder
@@ -39,18 +42,19 @@ var SliderComponent = /** @class */ (function () {
     }
     Object.defineProperty(SliderComponent.prototype, "value", {
         get: function () {
-            if (!this._value)
+            if (!this._values)
                 return 0;
-            if (!this._value.join)
-                return this._value;
-            return this._value.join(',');
+            if (this.multiple)
+                return this._values.slice().sort((function (a, b) { return a - b; })).join(',');
+            return this._values[0];
         },
         set: function (val) {
-            if (val !== this._value) {
-                this._value = val;
-                this.onChangeCallback(this._value);
+            val = ('' + val).split(',');
+            if (String(val) !== String(this._values)) {
+                this.setValues(val);
+                this.onChangeCallback(this._values);
                 this.change.emit({
-                    value: this.value,
+                    value: this._values,
                     percent: this.percent
                 });
             }
@@ -58,9 +62,12 @@ var SliderComponent = /** @class */ (function () {
         enumerable: true,
         configurable: true
     });
-    Object.defineProperty(SliderComponent.prototype, "isFilled", {
+    Object.defineProperty(SliderComponent.prototype, "percent", {
         get: function () {
-            return this.filled;
+            var pct = this._percents;
+            if (this.multiple)
+                return pct.join(',');
+            return '' + pct[0];
         },
         enumerable: true,
         configurable: true
@@ -79,23 +86,42 @@ var SliderComponent = /** @class */ (function () {
         enumerable: true,
         configurable: true
     });
-    Object.defineProperty(SliderComponent.prototype, "isActive", {
-        get: function () {
-            return this.active;
-        },
-        enumerable: true,
-        configurable: true
-    });
-    Object.defineProperty(SliderComponent.prototype, "percent", {
-        get: function () {
-            return Math.round(100 * (this.value - this.min) / (this.max - this.min));
-        },
-        enumerable: true,
-        configurable: true
-    });
+    SliderComponent.prototype.setValues = function (values) {
+        var _this = this;
+        this._values = values;
+        this._percents = values
+            .map(function (v) { return Math.max(_this.min, Math.min(_this.max, v)); })
+            .map(function (v) { return Math.round(100 * (v - _this.min) / (_this.max - _this.min)); });
+        this._thumbs = this._percents
+            .map(function (p) {
+            return {
+                left: "calc(" + p + "% - " + p / 100 + "em)"
+            };
+        });
+        if (this.filled) {
+            this._fill = this.getFill();
+        }
+        if (this.showTicks) {
+            this._ticks = this.getTicks();
+        }
+    };
+    SliderComponent.prototype.setActive = function (index, active) {
+        this._active[index] = active;
+    };
     SliderComponent.prototype.ngOnInit = function () {
         if (this.showTicks) {
-            this.count = this.getCount();
+            this._ticks = this.getTicks();
+        }
+    };
+    SliderComponent.prototype.setValue = function (val, index) {
+        if (this._values[index] !== val) {
+            this._values[index] = val;
+            this.setValues(this._values);
+            this.onChangeCallback(this.value);
+            this.change.emit({
+                value: this.value,
+                percent: this.percent
+            });
         }
     };
     SliderComponent.prototype.getCount = function () {
@@ -108,11 +134,27 @@ var SliderComponent = /** @class */ (function () {
         }
         return idxs;
     };
+    SliderComponent.prototype.getTicks = function () {
+        return this.getCount().map(function (p) {
+            return {
+                left: "calc(" + p + "% - " + (p / 100 - 0.5) + "em)"
+            };
+        });
+    };
     SliderComponent.prototype.getFill = function () {
         if (this.filled) {
-            var size = this.isHorizontal ? this.percent + "% 100%" : "100% " + this.percent + "%";
+            var percentMin = this.multiple ? Math.min.apply(Math, this._percents) : 0;
+            var percentMax = this.multiple ? Math.max.apply(Math, this._percents) : this._percents[0];
+            var width = percentMax - percentMin;
+            if (edge && this.multiple) {
+                return {
+                    left: "calc(" + percentMin + "% - " + (percentMin / 100 - 0.5) + "em)",
+                    'background-size': "calc(" + width + "% - " + width / 100 + "em) 100%"
+                };
+            }
             return {
-                'background-size': size
+                left: percentMin + "%",
+                'background-size': width + "% 100%"
             };
         }
     };
@@ -132,8 +174,9 @@ var SliderComponent = /** @class */ (function () {
         });
     };
     SliderComponent.prototype.writeValue = function (val) {
-        if (val !== this._value) {
-            this._value = val;
+        val = String(val).split(',');
+        if (String(val) !== String(this._values)) {
+            this.setValues(val);
         }
     };
     SliderComponent.prototype.registerOnChange = function (fn) {
@@ -141,6 +184,9 @@ var SliderComponent = /** @class */ (function () {
     };
     SliderComponent.prototype.registerOnTouched = function (fn) {
         this.onTouchedCallback = fn;
+    };
+    SliderComponent.prototype.trackIndex = function (index) {
+        return index;
     };
     __decorate([
         Input(),
@@ -163,13 +209,20 @@ var SliderComponent = /** @class */ (function () {
         __metadata("design:type", String)
     ], SliderComponent.prototype, "orientation", void 0);
     __decorate([
+        HostBinding('class.filled'),
         Input(),
         __metadata("design:type", Boolean)
     ], SliderComponent.prototype, "filled", void 0);
     __decorate([
+        HostBinding('class.multiple'),
         Input(),
         __metadata("design:type", Boolean)
     ], SliderComponent.prototype, "multiple", void 0);
+    __decorate([
+        HostBinding('class.disabled'),
+        Input(),
+        __metadata("design:type", Boolean)
+    ], SliderComponent.prototype, "disabled", void 0);
     __decorate([
         Input(),
         __metadata("design:type", Boolean)
@@ -179,14 +232,13 @@ var SliderComponent = /** @class */ (function () {
         __metadata("design:type", Number)
     ], SliderComponent.prototype, "tickStep", void 0);
     __decorate([
+        HostBinding('class.active'),
+        __metadata("design:type", Boolean)
+    ], SliderComponent.prototype, "active", void 0);
+    __decorate([
         Output(),
         __metadata("design:type", Object)
     ], SliderComponent.prototype, "change", void 0);
-    __decorate([
-        HostBinding('class.filled'),
-        __metadata("design:type", Boolean),
-        __metadata("design:paramtypes", [])
-    ], SliderComponent.prototype, "isFilled", null);
     __decorate([
         HostBinding('class.horizontal'),
         __metadata("design:type", Boolean),
@@ -197,11 +249,6 @@ var SliderComponent = /** @class */ (function () {
         __metadata("design:type", Boolean),
         __metadata("design:paramtypes", [])
     ], SliderComponent.prototype, "isVertical", null);
-    __decorate([
-        HostBinding('class.active'),
-        __metadata("design:type", Boolean),
-        __metadata("design:paramtypes", [])
-    ], SliderComponent.prototype, "isActive", null);
     __decorate([
         HostListener('mousedown', ['$event']),
         __metadata("design:type", Function),
@@ -217,7 +264,7 @@ var SliderComponent = /** @class */ (function () {
     SliderComponent = __decorate([
         Component({
             selector: 'ngx-slider',
-            template: "\n    <div class=\"slider-inner\">\n      <input\n        type=\"range\"\n        [id]=\"id\"\n        [attr.list]=\"id + '-list'\"\n        [attr.orientation]=\"orientation\"\n        [(ngModel)]=\"value\"\n        [min]=\"min\"\n        [max]=\"max\"\n        [multiple]=\"multiple\"\n        [step]=\"step\"\n        (input)=\"onChange($event)\"\n        (change)=\"onChange($event)\"\n      />\n      <span\n        *ngIf=\"filled\"\n        [ngStyle]=\"getFill()\"\n        class=\"fill-bar\">\n      </span>\n      <datalist\n        *ngIf=\"showTicks\"\n        [id]=\"id + '-list'\">\n        <option *ngFor=\"let i of count\">\n          {{i}}\n        </option>\n      </datalist>\n    </div>\n  ",
+            template: "\n    <div class=\"slider-inner\">\n      <div class=\"ticks-container\" *ngIf=\"showTicks\">\n        <div class=\"tick\" *ngFor=\"let s of _ticks\" [ngStyle]=\"s\">\n        </div>\n      </div>\n      <div class=\"inputs\">\n        <div class=\"slider-track\">\n        </div>\n        <span\n          *ngIf=\"filled\"\n          [ngStyle]=\"_fill\"\n          class=\"fill-bar\">\n        </span>\n        <ng-container *ngFor=\"let value of _values; let i = index; let odd = odd; trackBy: trackIndex\">\n          <input\n            type=\"range\"\n            [id]=\"id + '-' + i\"\n            [attr.list]=\"id + '-list'\"\n            [attr.orientation]=\"orientation\"\n            [class.odd]=\"odd\"\n            [class.active]=\"_active[i]\"\n            [ngModel]=\"value\"\n            (ngModelChange)=\"setValue($event, i)\"\n            [min]=\"min\"\n            [max]=\"max\"\n            [step]=\"step\"\n            [disabled]=\"disabled\"\n            (input)=\"onChange($event)\"\n            (change)=\"onChange($event)\"\n            (mouseenter)=\"setActive(i, true)\"\n            (mouseleave)=\"setActive(i, false)\"\n          />\n          <div class=\"slider-thumb\" [class.active]=\"_active[i]\" [ngStyle]=\"_thumbs[i]\" >\n          </div>\n        </ng-container>\n      </div>\n    </div>\n  ",
             encapsulation: ViewEncapsulation.None,
             styleUrls: ['./slider.component.css'],
             providers: [SLIDER_VALUE_ACCESSOR],