+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## HEAD (Unreleased)
 
+# 21.4.0 (2018-08-08)
+
+* Enhancement: Improvments to ngx-select (#186)
+* Enhancement: Various dropdown styles (#188)
+* Enhancement: Improved slider (#182)
+* Enhancement: Added icons (#191)
+* Bug: Improve invalid state for ngx-input (#189)
+
 # 21.3.0 (2018-07-10)
 
 * Enhancement: Added icons